create a component for SNS icons
