+++ src/styles/footer.module.scss
@@ -16,13 +16,6 @@
     align-items: center;
     justify-content: space-between;
     margin: 0 auto 2rem;
-    .sns_icon {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      border-radius: 0.5rem;
-      overflow: hidden;
-    }
   }
   .copyright {
     font-size: 1.2rem;
