+++ src/styles/sns.module.scss
@@ -1,0 +1,7 @@
+.icon {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 0.5rem;
+  overflow: hidden;
+}
